Updated selection to pass more information about location relative to the parent tag
Updated SelectionHandler and SelectionUtils to pass more information about the highlighted text's position relative to the parent tag. Selection positions are global to the whole document. If the document is modified, this information may not reference the desired text.. Relative positions could provide more stable information for documents that change over time. There are limitations to this solution as selections that span sections may be unstable or provide incorrect information.