Fix how annotation positions are matched to the dom
Match sections in textNodeProperties to the charOffsets [start, end]
There are separate if statements for the start and end positions because the boundary conditions are slightly different. This prevents preceding sections from being included if the end boundary touches the start of charOffsets[0]. This has the added benefit of having relations snap to the correct section.